@@ -6,10 +6,24 @@
     "lit-html": "~1.0.0-rc.2"
   },
   "devDependencies": {
+    "@types/copy-webpack-plugin": "^5.0.1",
+    "@types/webpack": "^4.41.13",
+    "autoprefixer": "^9.6.1",
+    "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "~4.6.0",
-    "pulito": "^4.1.1",
+    "css-loader": "^3.5.3",
+    "cssnano": "^3.10.0",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^4.3.0",
+    "mini-css-extract-plugin": "^0.9.0",
+    "node-sass": "^4.14.1",
+    "postcss-loader": "^2.1.6",
+    "sass-loader": "^7.3.1",
     "terser-webpack-plugin": "^1.2.2",
-    "webpack": "~4.28.0",
-    "webpack-cli": "^3.2.1"
+    "ts-node": "^8.10.1",
+    "typescript": "^3.9.2",
+    "webpack": "^4.40.2",
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-server": "^3.8.1"
   }
 }
